Move git service and it's configuration
Move git service and it's configuration into the new package `services.git`.
