+++ src/main/java/com/evosome/regit/services/git/properties/GitServiceProperties.java
@@ -1,4 +1,4 @@
-package com.evosome.regit.properties;
+package com.evosome.regit.services.git.properties;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -9,7 +9,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @Getter
 @Setter
 @Component
-@ConfigurationProperties("re-git")
+@ConfigurationProperties("git-service")
 public class GitServiceProperties {
 
     @Value("clone-path")
